Algorithm seeding "fix"
When using `partial` for to fix `rng` for reproducibility, if `int` seed is used, the RNG will be redefined at the same point every call. If instead a `Generator` object is passed to `rng`, its state is preserved between calls. Use the former if you want every call of the algorithm to be the same. Use the latter if you want the same sequence of distinct random evaluations.