Added PyTorch module wrapper for valid action enforcement
See `valid_wrapper` function in `learning.supervised.torch.Base.__init__`. Relevant for `StepTasking` environments. Invokes basic functionality of `StepTasking.infer_action_space` to generate a mask, which is to create a new module `forward` method. The modified network is "parallel", with the user-specified network in once path and the masking in another