Generalize `evaluate_algorithms_train` for multiple learners. Closes #8
Each iteration, problems (and solutions) populate local variables. For each learner (and iteration), the data is used to instantiate new `Dataset` generators; these generators are assigned to the `env.problem_gen` for each learner for training. This guarantees that the same problems are used to for each learner, for each iteration.

Also:
- Add base SL class to determine which learners need solutions
- Cleanup